+++ decidim-comments/spec/cells/decidim/comments/comment_activity_cell_spec.rb
@@ -25,6 +25,21 @@ module Decidim::Comments
         expect(html).to have_content(comment.body)
       end
 
+      context "when the comment has mentions" do
+        before do
+          body = "Comment mentioning some user, @#{comment.author.nickname}"
+          parsed_body = Decidim::ContentProcessor.parse(body, current_organization: comment.organization)
+          comment.body = parsed_body.rewrite
+          comment.save
+        end
+
+        it "correctly renders comments with mentions" do
+          html = cell("decidim/comments/comment_activity", action_log).call
+          expect(html).to have_no_content("gid://")
+          expect(html).to have_content("@#{comment.author.nickname}")
+        end
+      end
+
       context "when the commentable is missing" do
         before do
           comment.root_commentable.delete