+++ decidim-core/config/routes.rb
@@ -16,6 +16,14 @@ Decidim::Core::Engine.routes.draw do
                omniauth_callbacks: "decidim/devise/omniauth_registrations"
              }
 
+  devise_for :user_groups,
+             class_name: "Decidim::UserGroup",
+             module: :devise,
+             router_name: :decidim,
+             controllers: {
+               confirmations: "decidim/devise/confirmations"
+             }
+
   devise_scope :user do
     post "omniauth_registrations" => "devise/omniauth_registrations#create"
   end
@@ -71,6 +79,7 @@ Decidim::Core::Engine.routes.draw do
           post :demote
         end
       end
+      resource :email_confirmation, only: [:create], controller: "group_email_confirmations"
 
       member do
         delete :leave